requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector==2.2.9
mysql-connector-python==8.0.22
praw==7.1.0
prawcore==1.5.0
protobuf==3.14.0
pydotenv==0.0.7
pyTelegramBotAPI==3.7.4
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
